+++ src/components/SprigPlayground.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace putyourlightson\sprig\plugin\components;
 
 use Craft;
@@ -9,17 +10,20 @@ use yii\web\ForbiddenHttpException;
 
 class SprigPlayground extends Component
 {
-    public $variables = [];
+    /**
+     * @var array
+     */
+    public array $variables = [];
 
     /**
      * @inheritdoc
      */
-    public function init()
+    public function init(): void
     {
         parent::init();
 
         // Validate that user has permission to access the plugin in the CP
-        if (!Craft::$app->getUser()->checkPermission('accessPlugin-'.Sprig::$plugin->id)) {
+        if (!Craft::$app->getUser()->checkPermission('accessPlugin-' . Sprig::$plugin->id)) {
             throw new ForbiddenHttpException('Access denied.');
         }
     }
@@ -72,6 +76,6 @@ class SprigPlayground extends Component
     {
         $error = preg_replace('/in "__string_template__(.*?)"/', '', $error);
 
-        return '<h2 class="error">'.$error.'</h2>';
+        return '<h2 class="error">' . $error . '</h2>';
     }
 }